+++ frontend/components/ui/Inputs/InputFotos.tsx
